Optimize widget tree by using const constructors
In the given diff, we've used const constructors for widgets that don't change after the initial build.
This allows Flutter to optimize rebuilding widget tree. The optimization helps to skip creating a new Widget instance and directly use the existing instance from cache. This can improve the app's performance when these widgets are part of a larger widget tree that needs frequent rebuilding. We've applied it to various instances of widgets like PublicNotice, Row, DrawerHeader, SizedBox, Icon, and Text.