"Add 'provider' and 'nested' dependencies to project
The 'provider' package ('^6.0.5') is introduced to manage state and provide a better way of distributing and consuming app-wide data. This helps efficiently manage app state and optimize UI updates across the app.

Simultaneously, the 'nested' package ('^1.0.0') is added to handle nested widget structures more efficiently. It handles dependencies between different parts of the project such as the PostList, PostRepository, and PostViewModel."