Rearrange layout and improve readability in main.dart
The order of pages array in _MyHomePageState in main.dart was updated, moving PostList to the top. This was done to improve the logical sequence of page display, enhancing user experience. The AppBar title was also updated for more clarity, specifying that it is the admin page for 'Seoul Forest', instead of being a generic admin page.