Add public notice creation modal and related references
Added a new `CreatePublicNoticeModal` Dart file that builds a modal for creating public notices. It facilitates the selection of a country, city, and other options for public notices. The `first_page.dart` file was also updated to import this new modal and include it within the user interface. This change allows for an additional functionality where users can create and customize announcements through this modal. The creation of these public notices is now more user-friendly with this new interface.