Refactor layout structure and simplify scrolling
Refactored the structure of the main and first page layouts. Removed the horizontal scroll direction in SingleChildScrollView from the main.dart file, as it was unnecessary. Also, a container with a blue background color was introduced to the first page to enhance visual appeal. This improves application's UI and enhances user experience.